requirements.txt
# Backend dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
PyPDF2>=2.0.0
pdfplumber>=0.7.4
matplotlib>=3.4.3
fpdf2>=2.5.5
wikipedia-api>=0.5.4
requests>=2.26.0
transformers>=4.11.0
wikipedia>=1.4.0
duckduckgo-search>=2.9.0

# Frontend dependencies (for reference, to be installed in the frontend directory)
# react
# react-dom
# @mui/material
# @emotion/react
# @emotion/styled
# axios
